refactor: Move src/Sort.ts to src/Query/Sort.ts
