+++ tests/Sort.test.ts
@@ -5,7 +5,7 @@ import moment from 'moment';
 
 window.moment = moment;
 
-import { Sort } from '../src/Sort';
+import { Sort } from '../src/Query/Sort';
 import { resetSettings, updateSettings } from '../src/config/Settings';
 import { fromLine } from './TestHelpers';
